:truck: api: move uncancel certif course usecase to src
